Fix: 오븐 상세보기 기능 수정

+++ src/main/java/com/cookiepaper/repository/OvenRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OvenRepository extends JpaRepository<Oven, Integer> {
+public interface OvenRepository extends JpaRepository<Oven, Long> {
 
     Oven getByUsId(String usId);
